refactor(import): sort the accumulated helpers into files by mechanism
The iterations left declarations stranded where the edits happened rather
than where they belong. Move the no-cycle edge filters next to the graph
build that reads them, split the HasExport statement walk out of
export_map.go into export_lookup.go, return exportAssignmentReferencedIdentifier
to its only caller, and drop the ambientNamespace parameter that no call
site set.
