+++ internal/plugins/import/utils/export_test.go
@@ -1,5 +1,5 @@
 package utils
 
 // CompileModuleSettings gives external tests the compiled form of raw
-// `import/` settings, which production callers reach through IndexFor.
+// `import/` settings, which production callers reach through SettingsFor.
 var CompileModuleSettings = compileModuleSettings
